refactor: change how digitraffic codegen is configured
This makes room for other Graphql endpoints
