+++ site/src/lib/digitraffic/queries/live_trains.ts
@@ -1,5 +1,5 @@
-import { graphql } from '~/generated'
-import { SimpleTrainFragment } from '~/generated/graphql'
+import { graphql } from '~/generated/digitraffic'
+import { SimpleTrainFragment } from '~/generated/digitraffic/graphql'
 
 export const trains = graphql(`
   query trains(
