+++ site/src/lib/digitraffic/queries/single_train.ts
@@ -1,5 +1,5 @@
-import { graphql } from '~/generated'
-import { SimpleTrainFragment } from '~/generated/graphql'
+import { graphql } from '~/generated/digitraffic'
+import { SimpleTrainFragment } from '~/generated/digitraffic/graphql'
 
 export const singleTrain = graphql(`
   query singleTrain($departureDate: Date!, $trainNumber: Int!) {
